Remove Action::AssetTransaction
This patch flattens AssetTransactions. For now, all asset transactions
become a type of Action.

An it also introduces new name `tracker`. It's a bare hash, hash without
signature and approvals, of the transaction.

And it renames a InnerTransaction to ShardTransaction.
