+++ state/src/impls/test_helper.rs
@@ -21,9 +21,45 @@ use ctypes::ShardId;
 pub const NETWORK_ID: &str = "tc";
 pub const SHARD_ID: ShardId = 0;
 
+macro_rules! mint_asset {
+    ($output:expr, $metadata:expr) => {
+        $crate::ctypes::parcel::Action::MintAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            shard_id: $crate::impls::test_helper::SHARD_ID,
+            metadata: $metadata,
+            output: $output,
+            approver: None,
+            administrator: None,
+            approvals: vec![],
+        }
+    };
+    ($output:expr, $metadata:expr, approver: $approver:expr) => {
+        $crate::ctypes::parcel::Action::MintAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            shard_id: $crate::impls::test_helper::SHARD_ID,
+            metadata: $metadata,
+            output: $output,
+            approver: Some($approver),
+            administrator: None,
+            approvals: vec![],
+        }
+    };
+    ($output:expr, $metadata:expr, administrator: $admin:expr) => {
+        $crate::ctypes::parcel::Action::MintAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            shard_id: $crate::impls::test_helper::SHARD_ID,
+            metadata: $metadata,
+            output: $output,
+            approver: None,
+            administrator: Some($admin),
+            approvals: vec![],
+        }
+    };
+}
+
 macro_rules! asset_mint {
     ($output:expr, $metadata:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetMint {
+        $crate::ctypes::transaction::ShardTransaction::MintAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             shard_id: $crate::impls::test_helper::SHARD_ID,
             metadata: $metadata,
@@ -33,7 +69,7 @@ macro_rules! asset_mint {
         }
     };
     ($output:expr, $metadata:expr, approver: $approver:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetMint {
+        $crate::ctypes::transaction::ShardTransaction::MintAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             shard_id: $crate::impls::test_helper::SHARD_ID,
             metadata: $metadata,
@@ -43,7 +79,7 @@ macro_rules! asset_mint {
         }
     };
     ($output:expr, $metadata:expr, administrator: $admin:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetMint {
+        $crate::ctypes::transaction::ShardTransaction::MintAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             shard_id: $crate::impls::test_helper::SHARD_ID,
             metadata: $metadata,
@@ -75,9 +111,9 @@ macro_rules! asset_mint_output {
 }
 
 macro_rules! asset_out_point {
-    ($hash:expr, $index:expr, $asset_type:expr, $amount:expr) => {
+    ($tracker:expr, $index:expr, $asset_type:expr, $amount:expr) => {
         $crate::ctypes::transaction::AssetOutPoint {
-            transaction_hash: $hash,
+            transaction_hash: $tracker,
             index: $index,
             asset_type: $asset_type,
             amount: $amount,
@@ -136,9 +172,52 @@ macro_rules! asset_transfer_outputs {
     };
 }
 
+macro_rules! transfer_asset {
+    (inputs: $inputs:expr, $outputs:expr) => {
+        $crate::ctypes::parcel::Action::TransferAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            burns: Vec::new(),
+            inputs: $inputs,
+            outputs: $outputs,
+            orders: Vec::new(),
+            approvals: vec![],
+        }
+    };
+    (inputs: $inputs:expr, $outputs:expr, approvals: $approvals:expr) => {
+        $crate::ctypes::parcel::Action::TransferAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            burns: Vec::new(),
+            inputs: $inputs,
+            outputs: $outputs,
+            orders: Vec::new(),
+            approvals: $approvals,
+        }
+    };
+    (inputs: $inputs:expr, $outputs:expr, $orders:expr) => {
+        $crate::ctypes::parcel::Action::TransferAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            burns: Vec::new(),
+            inputs: $inputs,
+            outputs: $outputs,
+            orders: $orders,
+            approvals: vec![],
+        }
+    };
+    (burns: $burns:expr) => {
+        $crate::ctypes::parcel::Action::TransferAsset {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            burns: $burns,
+            inputs: Vec::new(),
+            outputs: Vec::new(),
+            orders: Vec::new(),
+            approvals: vec![],
+        }
+    };
+}
+
 macro_rules! asset_transfer {
     (inputs: $inputs:expr, $outputs:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetTransfer {
+        $crate::ctypes::transaction::ShardTransaction::TransferAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             burns: Vec::new(),
             inputs: $inputs,
@@ -147,7 +226,7 @@ macro_rules! asset_transfer {
         }
     };
     (inputs: $inputs:expr, $outputs:expr, $orders:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetTransfer {
+        $crate::ctypes::transaction::ShardTransaction::TransferAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             burns: Vec::new(),
             inputs: $inputs,
@@ -156,7 +235,7 @@ macro_rules! asset_transfer {
         }
     };
     (burns: $burns:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetTransfer {
+        $crate::ctypes::transaction::ShardTransaction::TransferAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             burns: $burns,
             inputs: Vec::new(),
@@ -198,7 +277,7 @@ macro_rules! order_on_transfer {
 
 macro_rules! asset_compose {
     ($metadata:expr, $inputs:expr, $outputs:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetCompose {
+        $crate::ctypes::transaction::ShardTransaction::ComposeAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             shard_id: $crate::impls::test_helper::SHARD_ID,
             metadata: $metadata,
@@ -212,7 +291,7 @@ macro_rules! asset_compose {
 
 macro_rules! asset_decompose {
     ($input:expr, $outputs:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetDecompose {
+        $crate::ctypes::transaction::ShardTransaction::DecomposeAsset {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             input: $input,
             outputs: $outputs,
@@ -232,7 +311,7 @@ macro_rules! asset_wrap_ccc_output {
 
 macro_rules! asset_wrap_ccc {
     ($parcel_hash:expr, $output:expr) => {
-        $crate::ctypes::transaction::InnerTransaction::AssetWrapCCC {
+        $crate::ctypes::transaction::ShardTransaction::WrapCCC {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             shard_id: $crate::impls::test_helper::SHARD_ID,
             parcel_hash: $parcel_hash,
@@ -241,9 +320,19 @@ macro_rules! asset_wrap_ccc {
     };
 }
 
+macro_rules! unwrap_ccc {
+    ($burn:expr) => {
+        $crate::ctypes::parcel::Action::UnwrapCCC {
+            network_id: $crate::impls::test_helper::NETWORK_ID.into(),
+            burn: $burn,
+            approvals: vec![],
+        }
+    };
+}
+
 macro_rules! asset_unwrap_ccc {
     ($burn:expr) => {
-        $crate::ctypes::transaction::Transaction::AssetUnwrapCCC {
+        $crate::ctypes::transaction::ShardTransaction::UnwrapCCC {
             network_id: $crate::impls::test_helper::NETWORK_ID.into(),
             burn: $burn,
         }
